% Auto-generated: do not edit by hand
\docType{package}
\name{dashTable-package}
\alias{dashTable}
\title{Core Interactive Table Component for 'Dash'}
\description{
An interactive table component designed for editing and exploring large datasets, 'dashDataTable' is rendered with standard, semantic HTML <table/> markup, which makes it accessible, responsive, and easy to style. This component was written from scratch in 'React.js' specifically for the 'Dash' community. Its API was designed to be ergonomic and its behaviour is completely customizable through its properties.
}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plotly.com>
}
