% Auto-generated: do not edit by hand
\docType{package}
\name{dashTable-package}
\alias{dashTable}
\title{Core Interactive Table Component for Dash
}
\description{
Dash DataTable is an interactive table component designed for designed for viewing, editing, and exploring large datasets. DataTable is rendered with standard, semantic HTML <table/> markup, which makes it accessible, responsive, and easy to style. This component was written from scratch in React.js specifically for the Dash community. Its API was designed to be ergonomic and its behaviour is completely customizable through its properties.

}
\author{
\strong{Maintainer}: Ryan Patrick Kyle <ryan@plot.ly> 
}
